requirements.txt
fastapi
uvicorn
pydantic
python-multipart
boto3
python-jose
passlib
python-dotenv
opencv-python
